+++ lib/node/pl-drivers/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @milaboratories/pl-drivers
 
+## 1.16.22
+
+### Patch Changes
+
+- Updated dependencies [3e33e3c]
+- Updated dependencies [4068d5c]
+- Updated dependencies [13aff26]
+- Updated dependencies [ddc7746]
+- Updated dependencies [37707e3]
+  - @milaboratories/pl-tree@1.15.0
+  - @milaboratories/pl-client@3.17.0
+
 ## 1.16.21
 
 ### Patch Changes